Refactor and improve readability
