+++ src/components/tree/Tree.tsx
@@ -4,17 +4,17 @@ import { useDebounce } from 'react-use';
 
 import { useFilters } from '@/hooks/useFilters';
 import { useToolbox } from '@/hooks/useToolbox';
-import { TreeCatalog } from '@/pages/MapViewer/components/Toolbox';
 import { Collection } from '@/typings/stac';
 import { fetchPathPartsFromUrl } from '@/utils/genericUtils';
 
 import { TreeHeader } from './TreeHeader';
 import { TreeNode } from './TreeNode';
-import { filterTree } from './utils';
+import { buildTree, filterTree } from './utils';
+
 import './Tree.scss';
 
 type TreeProps = {
-  treeData: TreeCatalog[]; // This is the flat list from the API
+  treeData: TreeCatalog[];
   expandedNodes: { [key: string]: boolean };
   setExpandedNodes: Dispatch<SetStateAction<{ [key: string]: boolean }>>;
 };
@@ -35,59 +35,6 @@ export const Tree = ({ treeData, expandedNodes, setExpandedNodes }: TreeProps) =
       filterTree(node, activeFilters),
     [activeFilters],
   );
-
-  const buildTree = async (flatCatalogs: TreeCatalog[]): Promise<TreeCatalog[]> => {
-    const idToCatalog: { [id: string]: TreeCatalog } = {};
-    const rootCatalogs: TreeCatalog[] = [];
-
-    flatCatalogs.forEach((catalog) => {
-      idToCatalog[catalog.id] = { ...catalog, catalogs: [], collections: [] };
-    });
-
-    for (const catalog of flatCatalogs) {
-      const parentId = getParentId(catalog);
-      if (parentId && parentId !== 'root' && idToCatalog[parentId]) {
-        idToCatalog[parentId].catalogs.push(idToCatalog[catalog.id]);
-      } else {
-        rootCatalogs.push(idToCatalog[catalog.id]);
-      }
-    }
-
-    await Promise.all(
-      flatCatalogs.map(async (catalog) => {
-        const currentCatalog = idToCatalog[catalog.id];
-        const childLinks = currentCatalog.links.filter((link) => link.rel === 'child');
-
-        await Promise.all(
-          childLinks.map(async (link) => {
-            if (link.href.includes('/collections/')) {
-              try {
-                const response = await fetch(link.href);
-                const collection = await response.json();
-                currentCatalog.collections.push(collection);
-              } catch (error) {
-                console.error(`Failed to fetch collection at ${link.href}:`, error);
-              }
-            }
-          }),
-        );
-      }),
-    );
-
-    return rootCatalogs;
-  };
-
-  const getParentId = (catalog: TreeCatalog): string | null => {
-    const parentLink = catalog.links.find((link) => link.rel === 'parent');
-    if (parentLink) {
-      const hrefParts = parentLink.href.split('/catalogs/');
-      if (hrefParts.length > 1) {
-        const pathParts = hrefParts[1].split('/');
-        return pathParts[pathParts.length - 1];
-      }
-    }
-    return null;
-  };
 
   useEffect(() => {
     const parseURL = () => {
@@ -123,7 +70,6 @@ export const Tree = ({ treeData, expandedNodes, setExpandedNodes }: TreeProps) =
             .map((catalog) => filter(catalog))
             .filter(Boolean) as TreeCatalog[];
           setFilteredTreeData(filteredData);
-          console.log('Filtered Tree Data:', filteredData);
         })();
       }
     },
@@ -136,7 +82,6 @@ export const Tree = ({ treeData, expandedNodes, setExpandedNodes }: TreeProps) =
     if (url) {
       const path = url.split('catalogs/')[1];
       const currentPath = window.location.pathname;
-      // Preserve the suffix (/map or /list)
       const suffixMatch = currentPath.match(/\/(map|list)$/);
       const suffix = suffixMatch ? suffixMatch[0] : '';
       const newPath = `${import.meta.env.VITE_BASE_PATH || ''}/catalogs/${path}${suffix}`;
