+++ src/context/ToolboxContext/index.tsx
@@ -1,4 +1,4 @@
-import React, { createContext, useEffect, useReducer } from 'react';
+import React, { createContext, useCallback, useEffect, useReducer } from 'react';
 
 import { useFilters } from '@/hooks/useFilters';
 import { getStacItems } from '@/services/stac';
@@ -14,7 +14,6 @@ const initialState: ToolboxState = {
     type: 'FeatureCollection',
     features: [],
   },
-
   selectedCollectionItem: null,
   isCollectionItemsPending: false,
 };
@@ -31,6 +30,8 @@ const reducer = (state: ToolboxState, action: ToolboxAction): ToolboxState => {
       return { ...state, selectedCollectionItem: action.payload };
     case 'SET_COLLECTION_ITEMS_PENDING':
       return { ...state, isCollectionItemsPending: action.payload };
+    default:
+      return state;
   }
 };
 
@@ -44,19 +45,41 @@ const ToolboxProvider: React.FC<ToolboxProviderProps> = ({ children }) => {
 
   const catalogPath = getCatalogueFromURL();
 
-  const setSelectedCollectionItems = (selectedCollectionItems: ExtendedFeatureCollection) => {
-    dispatch({
-      type: 'SET_SELECTED_COLLECTION_ITEMS',
-      payload: selectedCollectionItems,
-    });
-  };
+  const setSelectedCollectionItems = useCallback(
+    (selectedCollectionItems: ExtendedFeatureCollection) => {
+      dispatch({
+        type: 'SET_SELECTED_COLLECTION_ITEMS',
+        payload: selectedCollectionItems,
+      });
+    },
+    [],
+  );
 
-  const setCollectionItemsPending = (isPending: boolean) => {
+  const setCollectionItemsPending = useCallback((isPending: boolean) => {
     dispatch({
       type: 'SET_COLLECTION_ITEMS_PENDING',
       payload: isPending,
     });
-  };
+  }, []);
+
+  // Memoize the action functions using useCallback
+  const setActivePage = useCallback((activePage: string) => {
+    dispatch({ type: 'SET_ACTIVE_PAGE', payload: activePage });
+  }, []);
+
+  const setSelectedCollection = useCallback((selectedCollection: Collection) => {
+    dispatch({
+      type: 'SET_SELECTED_COLLECTION',
+      payload: selectedCollection,
+    });
+  }, []);
+
+  const setSelectedCollectionItem = useCallback((selectedCollectionItem: StacItem) => {
+    dispatch({
+      type: 'SET_SELECTED_COLLECTION_ITEM',
+      payload: selectedCollectionItem,
+    });
+  }, []);
 
   useEffect(() => {
     const fetchItems = async () => {
@@ -85,11 +108,11 @@ const ToolboxProvider: React.FC<ToolboxProviderProps> = ({ children }) => {
     activeFilters.temporal.start,
     catalogPath,
     state.selectedCollection,
+    setCollectionItemsPending,
+    setSelectedCollectionItems,
   ]);
 
-  // Inside of selectedCollectionItems there is 100 items.
-  // We only want to return 10 at a time, in relation to the active page the user is on. Ignore the resultsPerPage for now.
-  const returnResultsPage = () => {
+  const returnResultsPage = useCallback(() => {
     if (state.selectedCollectionItems?.features) {
       const returnResultsPageOutput = state.selectedCollectionItems.features.slice(
         (activeFilters.resultsPage - 1) * 10,
@@ -98,32 +121,15 @@ const ToolboxProvider: React.FC<ToolboxProviderProps> = ({ children }) => {
       return returnResultsPageOutput;
     }
     return [];
-  };
+  }, [activeFilters.resultsPage, state.selectedCollectionItems?.features]);
 
   const value = {
     state,
     actions: {
-      setActivePage: (activePage: string) => {
-        dispatch({ type: 'SET_ACTIVE_PAGE', payload: activePage });
-      },
-      setSelectedCollection: (selectedCollection: Collection) => {
-        dispatch({
-          type: 'SET_SELECTED_COLLECTION',
-          payload: selectedCollection,
-        });
-      },
-      setSelectedCollectionItems: (selectedCollectionItems: ExtendedFeatureCollection) => {
-        dispatch({
-          type: 'SET_SELECTED_COLLECTION_ITEMS',
-          payload: selectedCollectionItems,
-        });
-      },
-      setSelectedCollectionItem: (selectedCollectionItem: StacItem) => {
-        dispatch({
-          type: 'SET_SELECTED_COLLECTION_ITEM',
-          payload: selectedCollectionItem,
-        });
-      },
+      setActivePage,
+      setSelectedCollection,
+      setSelectedCollectionItems,
+      setSelectedCollectionItem,
       returnResultsPage,
     },
   };
