+++ src/utils/genericUtils.tsx
@@ -61,3 +61,11 @@ export const fetchPathPartsFromUrl = () => {
   }
   return catalogPathParts;
 };
+
+export const fetchData = async (url: string) => {
+  const response = await fetch(url);
+  if (!response.ok) {
+    throw new Error('Error fetching data');
+  }
+  return response.json();
+};
